% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_maf_triplet.R
\name{annotate_maf_triplet}
\alias{annotate_maf_triplet}
\title{Annotate MAF with triplet context}
\usage{
annotate_maf_triplet(
  maf,
  all_SNVs = TRUE,
  ref,
  alt,
  fastaPath,
  bsgenome_name,
  pyrimidine_collapse = FALSE
)
}
\arguments{
\item{maf}{MAF file (required columns: Reference_Allele, Tumor_Seq_Allele2, NCBI_Build)}

\item{all_SNVs}{If TRUE, all triplet sequences of single nucleotide
variants (SNVs) are returned without filtering for specific reference (ref) or
alternative (alt) alleles. (Default is TRUE).
When all_SNVs is TRUE, the ref and alt parameters are ignored.}

\item{ref}{Reference allele (Only relevant when all_SNVs is FALSE; otherwise, this parameter is ignored.)}

\item{alt}{Alternative allele (Only relevant when all_SNVs is FALSE; otherwise, this parameter is ignored.)}

\item{fastaPath}{The path to the genome FASTA file corresponding
to the specified genome build. This file is used to extract sequence context
when no BSgenome package is provided via bsgenome_name.
- On GSC systems: If not specified, the function attempts to automatically
infer the FASTA path from the GAMBL configuration.
- On local systems: A valid local path to a FASTA file must be provided.}

\item{bsgenome_name}{Specifies the name of a BSgenome data package
to be used for sequence extraction. This parameter overrides fastaPath if provided.
- Format: \verb{BSgenome.<organism>.<provider>.<assembly>[.<masked>]}.
- Example: \code{"BSgenome.Hsapiens.UCSC.hg38"} for the human UCSC genome build hg38.
- If a masked genome is required, use a name like \code{"BSgenome.Hsapiens.UCSC.hg38.masked"}.}

\item{pyrimidine_collapse}{estimates the mutation strand
using a pyrimidine collapse strategy:
- Reference alleles C or T are interpreted as mutations on the + strand.
- Reference alleles A or G are interpreted as mutations on the - strand.}
}
\value{
A data frame with two to three extra columns, in case
pyrimidine_collapse = FALSE, it will add triple sequence (seq) and the
strand (mutation_strand). When pyrimidine_collapse = T, another column
also will be added to the those extra columns which shows the mutation
(mutation)
}
\description{
Give triple sequence of mutated base with its adjacent bases (-1
and +1)
}
\details{
It gives the reference and alternative alleles and looks for the
rows of the data frame based on these values for + strand genes and their
complement alleles rows for - strand genes, then it can look for the adjacent
bases in that mutation position. Also, it can look for all the SNVs in
the MAF data frame and provide triple sequences for them (reverse complement
sequence for the - strand).
}
\examples{
maf <- GAMBLR.data::get_coding_ssm(these_sample_id = "DOHH-2")

annotate_maf_triplet(maf)
annotate_maf_triplet(maf, all_SNVs = FALSE, "C", "T")
annotate_maf_triplet(maf, ref = "C", alt = "T", pyrimidine_collapse = TRUE)

Use bsgenome_name as an alternative if you lack GSC access or a FASTA file.
annotate_maf_triplet(maf, bsgenome_name = "BSgenome.Hsapiens.UCSC.grch37")

}
