% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_browser_hub.R
\name{build_browser_hub}
\alias{build_browser_hub}
\title{Build UCSC browser track hub}
\usage{
build_browser_hub(
  regions_bed,
  these_samples_metadata,
  projection = "grch37",
  maf_data = NULL,
  local_web_host_dir = NULL,
  hub_dir = "my_hub",
  splitColumnName = "pathology",
  colour_column = "lymphgen",
  hub_name = basename(hub_dir),
  shortLabel = basename(hub_dir),
  longLabel = basename(hub_dir),
  contact_email,
  visibility = "squish",
  bigDataUrl_base = "https://github.com/morinlab/LLMPP/blob/main",
  bedToBigBed_path,
  these_seq_types = NULL,
  these_sample_ids = NULL
)
}
\arguments{
\item{regions_bed}{A BED-format table with the regions you want to retrieve SSMs
from. If not provided, the function defaults to aSHM regions,
either \code{GAMBLR.data::grch37_ashm_regions} or \verb{GAMBLR. data::hg38_ashm_regions}, depending on the provided projection.}

\item{these_samples_metadata}{Optional (but highly recommended) metadata
table. Only samples with seq_type "genome" or "capture" will be used.
Otherwise it will use all genome and capture samples from \code{get_gambl_metadata()}.}

\item{projection}{Obtain variants projected to this reference,
one of grch37 (default) or hg38.}

\item{maf_data}{Optional. A data frame of mutations in MAF format or maf_data object
(e.g. from \code{get_coding_ssm} or \code{get_ssm_by_sample}). It can be filtered by
regions and samples provided by parameters \code{regions_bed} and
\code{these_samples_metadata}, respectively.}

\item{local_web_host_dir}{Path to the directory that is a local copy of the
web host to be used. For example, if the hub should be hosted on GitHub,
\code{local_web_host_dir} should be the path to your local copy of the repository
directory. Default is NULL.}

\item{hub_dir}{Path to the directory (inside the web host directory) where you want
to store the files for your track hub. If this directory does not exist, it will be
created. Default is "my_hub".}

\item{splitColumnName}{A single string to indicate which column of these_samples_metadata
to use to split the MAF data into separate tracks. Default is "pathology".}

\item{colour_column}{A single string to use to colour the SSMs in the track, ideally not
the same as \code{splitColumnName}. Possible options are lymphgen (default),
pathology, genome_build (as in \code{these_samples_metadata},
can be different from projection), and mutation (corresponds to MAF Variant_Classification).}

\item{hub_name}{A string with the hub name (without spaces) to fill in the \code{hub}
field of the hub.txt file. Default is \code{basename(hub_dir)}.}

\item{shortLabel}{A string with the short hub label (maximum of 17 characters
recommended; spaces are allowed) to fill in the \code{shortLabel} field of the
hub.txt file. Default is \code{basename(hub_dir)}.}

\item{longLabel}{A string with the long hub label (maximum of 80 characters
recommended; spaces are allowed) to fill in the \code{longLabel} field of the
hub.txt file. Default is \code{basename(hub_dir)}.}

\item{contact_email}{Required parameter. A string with the contact email to fill
in the \code{email} field of the hub.txt file.}

\item{visibility}{A string that controls the track visibility mode. Possible
values are "pack", "dense", "full", or "squish" (default).}

\item{bigDataUrl_base}{A string with the base path of the public web location of the
tracks' data files, to fill in the \code{bigDataUrl} fields from the hub.txt file.
For example, if the hub is hosted on GitHub, \code{bigDataUrl_base} should be
something like "https://github.com/morinlab/LLMPP/blob/main" (the default).
See how the track's paths from the bigDataUrl fields are set in the \strong{Details}
section.}

\item{bedToBigBed_path}{Path to your local \code{bedToBigBed} UCSC tool. If missing,
\code{GAMBLR.helpers::check_config_value} is called internally and the \code{bedToBigBed}
path is obtained from the \code{config.yml} file saved in the current working
directory.}

\item{these_seq_types}{Deprecated. Inferred from these_samples_metadata.}

\item{these_sample_ids}{Deprecated. Inferred from these_samples_metadata.}
}
\value{
Nothing.
}
\description{
Create a directory that contains the track hub files. They are:
a hub.txt file (named as \verb{<projection>_hub.txt} and marked with \verb{useOneFile on}),
and a subdirectory (named as the projection in use) with the custom tracks to
be visualized in UCSC browser. The custom track \code{regions} contains the regions
from where SSMs were retrieved. All other custom tracks contain SSMs from samples
separated by the \code{splitColumnName} parameter, where each file is named according
to the value in \code{splitColumnName} that it refers to.
}
\details{
\code{build_browser_hub} will create a custom track file for each combination of
\code{these_seq_types} and \code{splitColumnName} (if mutations can be found). Custom
track files are named as \verb{<seq_type_value>_<splitColumnName_value>.bb}.

The \code{bigDataUrl} field of a track in the hub.txt file is defined in
the following way:

\if{html}{\out{<div class="sourceCode">}}\preformatted{file.path(bigDataUrl_base, hub_dir, projection,
          paste0(track_file_names[i], "?raw=true\\n"))
}\if{html}{\out{</div>}}

where \code{track_file_names[i]} is the custom track file name as shown above.
}
\examples{
\dontrun{
# create a track hub in LLMPP GitHub repo

library(GAMBLR.data)

local_web_host_dir = "~/repos/LLMPP"
hub_dir = "hubs/ashm_test"

my_meta = get_gambl_metadata() \%>\%
  filter(pathology \%in\% c("BL", "DLBCL", "FL")) \%>\%
  filter(seq_type \%in\% c("genome", "capture"))

build_browser_hub(
  these_samples_metadata = my_meta,
  projection = "grch37",
  local_web_host_dir = local_web_host_dir,
  hub_dir = hub_dir,
  splitColumnName = "pathology",
  longLabel = "Public aSHM mutations separated by pathologies",
  contact_email = "rdmorin@sfu.ca"
)
}

}
