% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ashm_count_matrix.R
\name{get_ashm_count_matrix}
\alias{get_ashm_count_matrix}
\title{Get ASHM Count Matrix.}
\usage{
get_ashm_count_matrix(
  regions_bed,
  maf_data,
  these_samples_metadata,
  this_seq_type = "genome",
  projection
)
}
\arguments{
\item{regions_bed}{A bed file with one row for each region.
The first three columns in this file MUST contain the Chromosome, Start and End position}

\item{maf_data}{Optionally provide a data frame in the MAF format, otherwise
either GAMBLR.data or GAMBLR.results will be used.}

\item{these_samples_metadata}{This is used to complete your matrix. All GAMBL
samples of the specified seq_type will be used by default. Provide a data frame with at least
sample_id for all samples if you are using non-GAMBL data.}

\item{this_seq_type}{The seq_type to return results for. Must be a single value. Only used if no
metadata is provided with these_samples_metadata.}

\item{projection}{The genome build we are working with}
}
\value{
A data frame with a row for every sample in these_samples_metadata and a column for every region in regions_bed
}
\description{
Prepare a matrix with one row per sample and one column per
region using a set of hypermutated regions.
}
\details{
Values are the number of mutations in that patient in the region.
}
\examples{


  DLBCL_genome_meta = get_gambl_metadata() \%>\% 
    dplyr::filter(pathology=="DLBCL")
#get ashm regions
some_regions = GAMBLR.utils::create_bed_data(
                              GAMBLR.data::grch37_ashm_regions,
                              fix_names = "concat",
                              concat_cols = c("gene","region"),
                              sep="-") \%>\%
  dplyr::filter(grepl("PAX5",name))

pax5_matrix <- get_ashm_count_matrix(
     regions_bed = some_regions,
     this_seq_type = "genome",
     these_samples_metadata = DLBCL_genome_meta
)
head(pax5_matrix)
colMeans(pax5_matrix)

}
