% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cn_segments.R
\name{get_cn_segments}
\alias{get_cn_segments}
\title{Get CN Segments.}
\usage{
get_cn_segments(
  these_samples_metadata,
  projection = "grch37",
  flavour = "combined",
  this_seq_type,
  max_CN = 20,
  verbose = FALSE
)
}
\arguments{
\item{these_samples_metadata}{User must provide a metadata table to restrict the data to the samples in your table.
The metadata also ensures the proper handling of duplicate sample_id across seq_types and ensures the
seq_type in the metadata faithfully represents the seq_type of the data}

\item{projection}{Desired genome coordinate system for returned CN segments. Default is "grch37".}

\item{flavour}{Specify what pipeline or source of data to use.}

\item{this_seq_type}{Deprecated.}

\item{max_CN}{Cap CN values at this value. Default is 20. This is to avoid extreme CN values dominating the colour scale in visualisations.
Available options are "combined" (for the merge of the best tool for each data type),
or one of "purecn_cnvkit", "purecn_denovo", or "battenberg".
Other thabn "combined", the other flavours are incomplete (e.g. limited to genome,
matched or capture samples).}

\item{verbose}{Set to TRUE for a chattier experience}

\item{fill_missing_with}{Deprecated.
This argument is no longer available as we are updating how GAMBL reports CN and log.ratio values.
For simplicity, dummy segments (if detected) are now dropped.}

\item{adjust_for_ploidy}{Deprecated.
This argument is no longer available as we are updating how GAMBL reports CN and log.ratio values.}
}
\value{
A data frame with CN segments for the specified region.
}
\description{
Retrieve all copy number segments from the GAMBL outputs
}
\details{
This merely loads and returns all the seg_data available for a projection
(genome build) and can assign a single value to dummy segments if they are
present/identified in the source file
}
\examples{
# Example for just exome/capture samples:
# Get metadata for just a few capture samples
capture_metadata <- suppressMessages(get_gambl_metadata()) \%>\%
  dplyr::filter(seq_type == "capture") \%>\%
  head()

# Load the copy number segments for capture samples using hg38 projection
capture_segments_hg38 <- get_cn_segments(
  these_samples_metadata = capture_metadata,
  projection = "hg38"
)
print(capture_segments_hg38)
# Note the default projection is "grch37"

genome_metadata <- suppressMessages(get_gambl_metadata()) \%>\%
  dplyr::filter(seq_type == "genome")

# Create a metadata table with a mix of seq_types
mixed_seq_type_meta <- dplyr::bind_rows(capture_metadata, genome_metadata)

## We can load the copy number segments for all samples across seq_types
## To differentiate samples that share a sample_id but differ in seq_type,
## a new column seg_seq_type is added to the output

all_seq_type_segs <- get_cn_segments(
  these_samples_metadata = mixed_seq_type_meta
)

dplyr::group_by(all_seq_type_segs,seg_seq_type) \%>\% dplyr::count()

}
