% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cn_states.R
\name{get_cn_states}
\alias{get_cn_states}
\title{Get CN States.}
\usage{
get_cn_states()
}
\value{
Copy number matrix with sample_id as rows and regions as columns.
}
\description{
Get a copy number matrix for all samples based on segmented data in the database.
This function is deprecated and has been replaced by
GAMBLR.utils::segmented_data_to_cn_matrix and GAMBLR.results::get_cn_segments
}
\details{
This function returns CN states for the specified regions using the CN data
from GAMBLR.results and (optionally) assumes regions with no data are diploid.
For how to determine/specify the coordinates of each region, refer to the
parameter descriptions and examples.
}
\examples{
\dontrun{
  get_cnv_and_ssm_status()
}
}
