% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fusions.R
\name{get_gene_fusions}
\alias{get_gene_fusions}
\title{Get gene fusions}
\usage{
get_gene_fusions(
  projection = "grch37",
  verbose = F,
  keep_genes,
  drop_genes,
  remove_adjacent_pairs = FALSE,
  these_samples_metadata,
  join_with_metadata = FALSE,
  harmonize_immunoglobulin_partners = FALSE
)
}
\value{
A data frame in a bedpe-like format
}
\description{
Retrieve fusions identified by FusionCatcher
}
