% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lymphgen.R
\name{get_lymphgen}
\alias{get_lymphgen}
\title{Get Lymphgen.}
\usage{
get_lymphgen(
  these_samples_metadata,
  flavour,
  lymphgen_file,
  keep_all_rows = FALSE,
  keep_original_columns = FALSE,
  streamlined = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{these_samples_metadata}{A metadata table to auto-subset the data to samples in that table before returning.}

\item{flavour}{Lymphgen flavour.}

\item{lymphgen_file}{Path to lymphgen file.}

\item{keep_all_rows}{Boolean parameter, default is FALSE.}

\item{keep_original_columns}{Boolean parameter, default is FALSE.}

\item{streamlined}{Boolean, set to true to get just a data frame with one column for sample_id and one for LymphGen class}

\item{verbose}{Boolean, set to TRUE to print informational messages. Useful for debugging. Default is FALSE}
}
\value{
If run with A list of data frames with the following names:
lymphgen (a data frame containing the tidy LymphGen output),
features (a binary matrix indicating which patients had each feature),
feature_annotation (a data frame with one row per LymphGen feature reduced to gene or arm, for arm-level events and summary statistics for the feature across the cohort),
features_long (a data frame with one row per LymphGen feature/patient event),
sample_annotation (a data frame with one row per sample and columns indicating the number of features for each LymphGen class in that sample)
}
\description{
Get a specific flavour of LymphGen from the main GAMBL outputs.
}
\details{
Get a specific flavour of LymphGen from the main GAMBL outputs and tidy the composites.
Optionally return a matrix of features instead
}
\examples{
my_meta = get_gambl_metadata()
lymphgen_all <- get_lymphgen(
           flavour = "no_cnvs.no_sv.with_A53", 
           these = my_meta,
            keep_original_columns = TRUE
        )
head(lymphgen_all$features[,c(1:14)])

head(lymphgen_all$lymphgen[,c(1:14)])

}
