% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_cn_segments.R
\name{get_sample_cn_segments}
\alias{get_sample_cn_segments}
\title{Get Sample CN Segments.}
\usage{
get_sample_cn_segments(
  these_sample_ids = NULL,
  these_samples_metadata = NULL,
  from_flatfile = TRUE,
  projection = "grch37",
  this_seq_type = "genome",
  with_chr_prefix = FALSE,
  streamlined = FALSE
)
}
\value{
A data frame of segments for a specific or multiple sample ID(s).
}
\description{
Get all segments for a single (or multiple) sample_id(s).
}
\details{
Deprecated. See \link{get_cn_segments}.
}
\keyword{internal}
