% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ssm_by_region.R
\name{get_ssm_by_region}
\alias{get_ssm_by_region}
\title{Get SSM By Region.}
\usage{
get_ssm_by_region(
  chromosome,
  qstart,
  qend,
  region = "",
  these_samples_metadata,
  maf_data,
  streamlined = FALSE,
  basic_columns = TRUE,
  tool_name = "slms-3",
  augmented = TRUE,
  projection = "grch37",
  min_read_support = 3,
  verbose = FALSE,
  these_sample_ids,
  this_seq_type
)
}
\arguments{
\item{chromosome}{The chromosome you are restricting to (with or without a chr prefix).}

\item{qstart}{Query start coordinate of the region of interest.}

\item{qend}{Query end coordinate of the region of interest.}

\item{region}{Full region definition specified as a character vector e.g. "chrX:1234-5678"
instead of specifying chromosome, start and end separately.}

\item{these_samples_metadata}{Optional metadata table.
If provided, it will return SSM calls for the only the samples in the metadata table.
Otherwise it will use all samples from \code{get_gambl_metadata()} of the appropriate seq_types.}

\item{maf_data}{Use an already loaded MAF data frame.
If you would like all columns of this input maf returned,
set \code{streamlined = FALSE} and \code{basic_columns = FALSE}.
Otherwise the first 45 columns will be returned.}

\item{streamlined}{If TRUE, only 3 columns will be returned:
start, sample_id, and region in the format "chr:start-end". Default is FALSE.
Note: if this parameter is TRUE, the function will disregard anything specified with \code{basic_columns}.}

\item{basic_columns}{Parameter to be used when streamlined is FALSE.
Set this parameter to TRUE (default) to return a MAF with the standard 45 columns.
Set to FALSE to return a MAF with all columns (116).
If you provided \code{maf_data} with more than 45 columns, set to FALSE to return all columns of
\code{maf_data}, otherwise it will return the first 45.}

\item{tool_name}{Accepts either "slms-3" (default) or "strelka2"
(forces \code{streamlined=TRUE}) to indicate which variant caller to use. Note: strelka2 will force
\code{augmented=FALSE} as data is not available in that case. Formerly called "mode".}

\item{augmented}{Default: TRUE. Setting to FALSE will subset from a pre-merged MAF.
Obtaining variants from the original MAFs for each sample would not be computationally efficient here.
Provide to \code{maf_data} the results of \code{get_ssm_by_samples} if those are needed.}

\item{projection}{Obtain variants projected to this reference, one of grch37 (default) or hg38.}

\item{min_read_support}{Only returns variants with at least this many reads in t_alt_count
(for cleaning up augmented MAFs). Default: 3.}

\item{verbose}{Boolean parameter set to FALSE per default.}

\item{these_sample_ids}{Deprecated. Inferred from these_samples_metadata}

\item{this_seq_type}{Deprecated. Inferred from these_samples_metadata}
}
\value{
A data frame of variants in 3 column format or in MAF-like format (one row per mutation).
}
\description{
Retrieve all SSMs from the GAMBL database within a single genomic coordinate range.
}
\details{
This function lets the user specify a region of interest
for obtaining SSM calls within that region. In most cases, you
should be using \link{get_ssm_by_regions} instead.
There are multiple ways a region can be specified. For example,
the user can provide the full region in a "region" format (chr:start-end)
to the \code{region} parameter.
Or, the user can provide chromosome, start and end coordinates individually
with \code{chr}, \code{start}, and \code{end} parameters.
}
\examples{
#basic usage
genomes = get_gambl_metadata() \%>\%
             dplyr::filter(seq_type \%in\% "genome")
my_mutations = GAMBLR.results:::get_ssm_by_region(
            region = "chr8:128723128-128774067",
            these_samples_metadata = genomes)


#keep all 116 columns in the read MAF
bcl2_all_details = GAMBLR.results:::get_ssm_by_region(
            region = "chr18:60796500-60988073",
            these_samples_metadata = genomes,
            basic_columns = FALSE)
}
\keyword{internal}
