% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_seq_types.R
\name{link_seq_types}
\alias{link_seq_types}
\title{Link seq type IDs}
\usage{
link_seq_types(
  these_sample_ids,
  given_seq_type,
  desired_seq_type,
  intermediary
)
}
\arguments{
\item{these_sample_ids}{A vector of sample IDs of the same seq type specified in \code{given_seq_type}.}

\item{given_seq_type}{A single string specifying the seq type of \code{these_sample_ids}. Possible
values are one of "genome", "capture", or "mrna".}

\item{desired_seq_type}{A single string specifying the seq type of the desired sample IDs.
Possible values are one of "genome", "capture", "mrna", or "any". If "any", sample IDs
of any other seq type that is not \code{given_seq_type} will be returned.}

\item{intermediary}{A single string of the column name in the metadata which will act as a
intermediary to link the given sample IDs with the desired sample IDs.}
}
\value{
A data frame with the given and desired sample IDs, values of the feature specified in
the \code{intermediary} parameter, and the \code{multi_desired_id} column.
}
\description{
Link sample IDs of a seq type to sample IDs of a different seq type that share a same
feature, \emph{e.g.} same \code{biopsy_id}.
}
\details{
If there is more than one sample id of the desired seq type associated with a same value
in \code{intermediary}, the \code{multi_desired_id} column of the output table will show 1 instead of 0.
}
\examples{
link_seq_types(
  these_sample_ids = c("00-14595_tumorA", "04-21622_tumorB", "01-20774T"), 
  given_seq_type = "genome",
  desired_seq_type = "mrna", 
  intermediary = "biopsy_id"
)

}
