% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gambl_metadata.R
\name{og_get_gambl_metadata}
\alias{og_get_gambl_metadata}
\title{Get GAMBL metadata.}
\usage{
og_get_gambl_metadata(
  seq_type_filter = "genome",
  tissue_status_filter = "tumour",
  case_set,
  remove_benchmarking = TRUE,
  with_outcomes = TRUE,
  from_flatfile = TRUE,
  sample_flatfile,
  biopsy_flatfile,
  only_available = TRUE,
  seq_type_priority = "genome"
)
}
\arguments{
\item{seq_type_filter}{Filtering criteria (default: all genomes).}

\item{tissue_status_filter}{Filtering criteria for tissue status. Possible values are "tumour" (the default) or "normal".}

\item{case_set}{Optional short name for a pre-defined set of cases avoiding any embargoed cases (current options: 'BLGSP-study', 'FL-study', 'DLBCL-study', 'FL-DLBCL-study', 'FL-DLBCL-all', 'DLBCL-unembargoed', 'BL-DLBCL-manuscript', 'MCL','MCL-CLL').}

\item{remove_benchmarking}{By default the FFPE benchmarking duplicate samples will be dropped.}

\item{with_outcomes}{Optionally join to gambl outcome data.}

\item{from_flatfile}{New default is to use the metadata in the flat-files from your clone of the repo. Can be overridden to use the database.}

\item{sample_flatfile}{Optionally provide the full path to a sample table to use instead of the default.}

\item{biopsy_flatfile}{Optionally provide the full path to a biopsy table to use instead of the default.}

\item{only_available}{If TRUE, will remove samples with FALSE or NA in the bam_available column (default: TRUE).}

\item{seq_type_priority}{For duplicate sample_id with different seq_type available, the metadata will prioritize this seq_type and drop the others. Possible values are "genome" or "capture".}
}
\value{
A data frame with metadata for each biopsy in GAMBL

\describe{
\item{compression}{Format of the original data used as input for our analysis pipelines (cram, bam or fastq)}
\item{bam_available}{Whether or not this file was available when last checked.}
\item{patient_id}{The anonymized unique identifier for this patient. For BC samples, this will be Res ID.}
\item{sample_id}{A unique identifier for the sample analyzed.}
\item{seq_type}{The assay type used to produce this data (one of "genome","capture, "mrna", "promethION")}
\item{genome_build}{The name of the genome reference the data were aligned to.}
\item{tissue_status}{Whether the sample was atumour or normal.}
\item{cohort}{Name for a group of samples that were added together (usually from a single study), often in the format {pathology}_{cohort_descriptor}.}
\item{library_id}{The unique identifier for the sequencing library.}
\item{pathology}{The diagnosis or pathology for the sample}
\item{time_point}{Timing of biopsy in increasing alphabetical order (A = diagnosis, B = first relapse etc)}
\item{protocol}{General protocol for library construction. e.g. "Ribodepletion", "PolyA", or "Genome"}
\item{ffpe_or_frozen}{Whether the nucleic acids were extracted from a frozen or FFPE sample}
\item{read_length}{The length of reads (required for RNA-seq libraries)}
\item{strandedness}{Whether the RNA-seq librayr construction was strand-specific and, if so, which strand. Required for RNAseq; "positive", "negative", or "unstranded")}
\item{seq_source_type}{Required for RNAseq. Usually the same value as ffpe_or_frozen but sometimes immunotube or sorted cells}
\item{data_path}{Symbolic link to the bam or cram file (not usually relevant for GAMBLR)}
\item{link_name}{Standardized naming for symbolic link (not usually relevant for GAMBLR)}
\item{data_path}{Symbolic link to the fastq file (not usually relevant for GAMBLR)}
\item{fastq_link_name}{Standardized naming for symbolic link for FASTQ file, if used (not usually relevant for GAMBLR)}
\item{unix_group}{Whether a source is external and restricted by data access agreements (icgc_dart) or internal (gambl)}
\item{COO_consensus}{TODO}
\item{DHITsig_consensus}{TODO}
\item{COO_PRPS_class}{TODO}
\item{DHITsig_PRPS_class}{TODO}
\item{DLBCL90_dlbcl_call}{TODO}
\item{DLBCL90_dhitsig_call}{TODO}
\item{res_id}{duplicate of sample_id for local samples and NA otherwise}
\item{DLBCL90_code_set}{Code set used for DLBCL90 call. One of DLBCL90 DLBCL90v2 DLBCL90v3}
\item{DLBCL90_dlbcl_score}{TODO}
\item{DLBCL90_pmbl_score}{TODO}
\item{DLBCL90_pmbl_call}{TODO}
\item{DLBCL90_dhitsig_score}{TODO}
\item{myc_ba}{Result from breakapart FISH for MYC locus}
\item{myc_cn}{Result from copy number FISH for MYC locus}
\item{bcl2_ba}{Result from breakapart FISH for BCL2 locus}
\item{bcl2_cn}{Result from copy number FISH for BCL2 locus}
\item{bcl6_ba}{Result from breakapart FISH for BCL6 locus}
\item{bcl6_cn}{Result from copy number FISH for BCL6 locus}
\item{time_since_diagnosis_years}{TODO}
\item{relapse_timing}{TODO}
\item{dtbx}{TODO. OR REMOVE?}
\item{dtdx}{TODO. OR REMOVE?}
\item{lymphgen_no_cnv}{TODO}
\item{lymphgen_with_cnv}{TODO}
\item{lymphgen_cnv_noA53}{TODO}
\item{lymphgen_wright}{The LymphGen call for this sample from Wright et all (if applicable)}
\item{fl_grade}{TODO}
\item{capture_frozen_sample_id}{TODO}
\item{capture_FFPE_sample_id}{TODO}
\item{capture_unknown_sample_id}{TODO}
\item{genome_frozen_sample_id}{TODO}
\item{genome_ctDNA_sample_id}{TODO}
\item{genome_FFPE_sample_id}{TODO}
\item{mrna_PolyA_frozen_sample_id}{TODO}
\item{mrna_Ribodepletion_frozen_sample_id}{TODO}
\item{mrna_Ribodepletion_frozen_sample_id}{TODO}
\item{XXX_cohort}{Cohort name for batch effect correction(?)}
\item{transformation}{TODO}
\item{relapse}{TODO}
\item{ighv_mutation_original}{TODO}
\item{normal_sample_id}{TODO}
\item{pairing_status}{TODO}
\item{ICGC_ID}{TODO}
\item{ICGC_XXX}{metadata value for ICGC cohort inferred from external metadata}
\item{detailed_pathology}{TODO}
\item{COO_final}{TODO}
\item{consensus_pathology}{TODO}
\item{lymphgen}{TODO}
\item{Tumor_Sample_Barcode}{Duplicate of sample_id for simplifying joins to MAF data frames}
\item{consensus_coo_dhitsig}{TODO}
\item{pathology_rank}{Numeric rank for consistent ordering of samples by pathology}
\item{lymphgen_rank}{Numeric rank for consistent ordering of samples by LymphGen}
\item{hiv_status}{TODO}
\item{CODE_XXX}{Event-free status at last follow-up for overall survival (OS), progression-free survival (PFS) etc. 0 = no event/censored. 1 = event}
\item{XXX_YEARS}{Time, in years, from diagnosis to last follow-up for overall survival (OS), progression-free survival (PFS) }
\item{alive}{Theoretically redundant with CODE_OS}
\item{is_adult}{Adult or pediatric at diagnosis. One of "Adult" for adults and "Pediatric" otherwise}
\item{age_group}{Adult_BL or Pediatric_BL or Other, specific to the BLGSP study}
\item{age}{patient age at diagnosis}
\item{sex}{The biological sex of the patient, if available. Allowable options: M, F, NA}
\item{tx_primary}{TODO}
\item{cause_of_death}{TODO}
}
}
\description{
Return metadata for a selection of samples.
}
\details{
This function returns metadata for GAMBL samples. Options for subset and filter the returned data are available.
For more information on how to use this function with different filtering criteria, refer to the parameter descriptions,
examples and vignettes. Embargoed cases (current options: 'BLGSP-study', 'FL-study', 'DLBCL-study', 'FL-DLBCL-study', 'FL-DLBCL-all', 'DLBCL-unembargoed', 'BL-DLBCL-manuscript', 'MCL','MCL-CLL')
}
\examples{
#basic usage
my_metadata = get_gambl_metadata()

#use pre-defined custom sample sets
only_blgsp_metadata = get_gambl_metadata(case_set = "BLGSP-study")

#override default filters and request metadata for samples other than tumour genomes,
#e.g. also get the normals
tumour_and_normal_metadata = get_gambl_metadata(tissue_status_filter = c('tumour','normal'))

non_redundant_genome_and_capture = get_gambl_metadata(seq_type_filter = c('genome', 'capture'),
                                                       seq_type_priority = "genome")
                                                       
absolutely_everything = get_gambl_metadata(seq_type_filter = c('genome', 'capture','mrna'), tissue_status_filter=c('tumour','normal'))

}
