% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_lymhoplot.R
\name{pretty_lymphoplot}
\alias{pretty_lymphoplot}
\title{LymphGen feature visualization.}
\usage{
pretty_lymphoplot(
  these_samples_metadata,
  flavour = "with_cnvs.with_sv.with_A53",
  lymphgen_all,
  with_A53 = FALSE,
  show_side_annotation = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{these_samples_metadata}{A data frame containing the metadata for your samples (LymphGen result will be subset to the sample_id in this data frame).}

\item{flavour}{Required if the output of get_lymphgen has not been provided.}

\item{lymphgen_all}{The list returned by get_lymphgen.}

\item{with_A53}{Set to FALSE if you are using LymphGen results without A53. Recommended option.}

\item{show_side_annotation}{Set to TRUE if you want the default ComplexHeatmap annotations on the sides of your heatmap. Default is FALSE.}

\item{verbose}{Set to TRUE to print helpful messages. Useful for debugging. Default is FALSE (not verbose).}
}
\value{
A ComplexHeatmap object
}
\description{
Make a heatmap showing the frequency of LymphGen features across a cohort of samples.
}
\examples{
meta_df = get_gambl_metadata(seq_type_filter = c("genome", "capture")) \%>\%
     dplyr::filter(pathology == "DLBCL")
pretty_lymphoplot(meta_df, show_side_annotation = TRUE)
pretty_lymphoplot(
     meta_df,
     show_side_annotation = TRUE,
     flavour = "with_cnvs.with_sv.no_A53"
)

}
