% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_maf_triplet.R
\name{annotate_maf_triplet}
\alias{annotate_maf_triplet}
\title{Annotate MAF with triplet context}
\usage{
annotate_maf_triplet(
  maf,
  all_SNVs = TRUE,
  ref,
  alt,
  genome_build,
  fastaPath,
  pyrimidine_collapse = FALSE
)
}
\arguments{
\item{maf}{MAF file (required columns: Reference_Allele, Tumor_Seq_Allele2)}

\item{all_SNVs}{To give us all the triplet sequences of SNVs and not
specifying any specific ref and alt alleles (default is TRUE)}

\item{ref}{Reference allele}

\item{alt}{Alternative allele}

\item{genome_build}{The genome build for the variants you are
working with (default is to infer it from the MAF)}

\item{fastaPath}{Can be a path to a FASTA file on a disk. When on GSC,
this is first attempted to be inferred from the gambl reference through
path specified in config. Local files are also accepted as value here.}

\item{pyrimidine_collapse}{Estimate mutation_strand and}
}
\value{
A data frame with two to three extra columns, in case
pyrimidine_collapse = FALSE, it will add triple sequence (seq) and the
strand (mutation_strand). When pyrimidine_collapse = T, another column
also will be added to the those extra columns which shows the mutation
(mutation)
}
\description{
Give triple sequence of mutated base with its adjacent bases (-1
and +1)
}
\details{
It gives the reference and alternative alleles and looks for the
rows of the data frame based on these values for + strand genes and their
complement alleles rows for - strand genes, then it can look for the adjacent
bases in that mutation position. Also, it can look for all the SNVs in
the MAF data frame and provide triple sequences for them (reverse complement
sequence for the - strand).
}
\examples{
maf <- GAMBLR.open::get_coding_ssm(projection="hg38") \%>\% head(n=500)
# peek at the data
dplyr::select(maf,1:12) \%>\% head()

maf_anno <- annotate_maf_triplet(maf)
dplyr::select(maf_anno,1:12,seq) \%>\% head()
# Each mutation is now associated with it's sequence context in the
# reference genome in a column named seq
\dontrun{
  annotate_maf_triplet(maf, all_SNVs = FALSE, "C", "T")
  annotate_maf_triplet(maf, ref = "C", alt = "T", pyrimidine_collapse = TRUE)
}
}
