% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAMBLR.results-package.R
\docType{package}
\name{GAMBLR.results-package}
\alias{GAMBLR.results}
\alias{GAMBLR.results-package}
\title{GAMBLR.results: Access results of the Genomic Analysis of Mature B-cell Lymphomas}
\description{
This package is part of the Genomic Analysis of Mature B-cell Lymphomas (GAMBL) project developed by the Morin Lab.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://morinlab.github.io/GAMBLR.results/}
}

}
\author{
\strong{Maintainer}: Kostiantyn Dreval \email{kdreval@sfu.ca} (\href{https://orcid.org/0000-0002-6214-2843}{ORCID})

Authors:
\itemize{
  \item Ryan Morin \email{rdmorin@sfu.ca} (\href{https://orcid.org/0000-0003-2932-7800}{ORCID})
  \item Vladimir Souza \email{vsouza@bcgsc.ca}
}

}
\keyword{internal}
