% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAMBLR.results-package.R
\docType{package}
\name{GAMBLR.results-package}
\alias{GAMBLR.results}
\alias{GAMBLR.results-package}
\title{GAMBLR.results: Access results of the Genomic Analysis of Mature B-cell Lymphomas}
\description{
This package is part of the Genomic Analysis of Mature B-cell Lymphomas (GAMBL) project developed by the Morin Lab.
}
\author{
\strong{Maintainer}: Vladimir Souza \email{vsouza@bcgsc.ca}

}
\keyword{internal}
