% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coding_ssm.R
\name{get_coding_ssm}
\alias{get_coding_ssm}
\title{Get Coding SSM.}
\usage{
get_coding_ssm(
  these_sample_ids = NULL,
  these_samples_metadata = NULL,
  force_unmatched_samples,
  projection = "grch37",
  this_seq_type = "genome",
  basic_columns = TRUE,
  maf_cols = NULL,
  augmented = TRUE,
  min_read_support = 3,
  groups = c("gambl", "icgc_dart"),
  include_silent = TRUE,
  engine,
  verbose = TRUE,
  limit_cohort,
  exclude_cohort,
  limit_pathology,
  limit_samples,
  from_flatfile
)
}
\arguments{
\item{these_sample_ids}{Optional, restrict the returned maf to a set of sample IDs (this parameter is replacing the \code{limit_samples} parameter).}

\item{these_samples_metadata}{Supply a metadata table to auto-subset the data to samples in that table before returning.}

\item{force_unmatched_samples}{Optional argument for forcing unmatched samples, using \link{get_ssm_by_samples}.}

\item{projection}{Reference genome build for the coordinates in the MAF file. The default is hg19 genome build.}

\item{this_seq_type}{The seq_type you want back, default is genome.}

\item{basic_columns}{Set to FALSE to override the default behavior of returning only the first 45 columns of MAF data.}

\item{maf_cols}{if basic_columns is set to FALSE, the user can specify what columns to be returned within the MAF. This parameter can either be a vector of indexes (integer) or a vector of characters (matching columns in MAF).}

\item{augmented}{default: TRUE. Set to FALSE if you instead want the original MAF from each sample for multi-sample patients instead of the augmented MAF}

\item{min_read_support}{Only returns variants with at least this many reads in t_alt_count (for cleaning up augmented MAFs)}

\item{groups}{Deprecated. Use either these_samples_metadata or these_sample_ids instead}

\item{include_silent}{Logical parameter indicating whether to include silent mutations into coding mutations. Default is TRUE.}

\item{engine}{Deprecated. Ignored}

\item{verbose}{Controls the "verboseness" of this function (and internally called helpers).}

\item{limit_cohort}{Deprecated. Use either these_samples_metadata or these_sample_ids instead}

\item{exclude_cohort}{Deprecated. Use either these_samples_metadata or these_sample_ids instead}

\item{limit_pathology}{Deprecated. Use either these_samples_metadata or these_sample_ids instead}

\item{limit_samples}{Deprecated. Use either these_samples_metadata or these_sample_ids instead}

\item{from_flatfile}{Deprecated. Ignored}
}
\value{
A data frame containing all the MAF data columns (one row per mutation).
}
\description{
Retrieve all coding SSMs from one seq_type in GAMBL in MAF-like format.
}
\details{
Effectively retrieve coding SSM calls. Multiple filtering parameters are available for this function.
For more information on how to implement the filtering parameters, refer to the parameter descriptions as well as examples in the vignettes.
Is this function not what you are looking for? Try one of the following, similar, functions; \link{get_coding_ssm_status},
\link{get_ssm_by_patients}, \link{get_ssm_by_sample}, \link{get_ssm_by_samples}, \link{get_ssm_by_region}, \link{get_ssm_by_regions}
}
\examples{
#basic usage
maf_data = get_coding_ssm(limit_cohort = c("BL_ICGC"))

maf_data = get_coding_ssm(limit_samples = "HTMCP-01-06-00485-01A-01D")

}
