% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gambl_metadata2.R
\name{get_gambl_metadata}
\alias{get_gambl_metadata}
\title{Get GAMBL metadata.}
\usage{
get_gambl_metadata(
  dna_seq_type_priority = "genome",
  capture_protocol_priority = "Exome",
  dna_preservation_priority = "frozen",
  exome_capture_space_priority = c("agilent-sureselect-human-all-exon-v7",
    "agilent-sureselect-V5-plus-utr", "idt-xgen-v2-grch37", "exome-utr-grch37",
    "exome-utr-grch38", "none"),
  mrna_collapse_redundancy = TRUE,
  collapse_redundancy = TRUE,
  also_normals = FALSE,
  include_unavailable = FALSE,
  verbose = FALSE,
  invert = FALSE,
  exclude = "promethION",
  min_corrected_cov = 25,
  ...
)
}
\arguments{
\item{dna_seq_type_priority}{The default is "genome" and the only other option is "capture". For duplicate biopsy_id/patient combinations with different seq_type available, prioritize this seq_type and drop the others.}

\item{capture_protocol_priority}{For duplicate biopsy_id/patient combinations with different seq_type available, prioritize this seq_type and drop the others.}

\item{dna_preservation_priority}{Which to prioritize between FFPE and frozen samples from the same biopsy (default: "frozen")}

\item{exome_capture_space_priority}{A vector specifying how to prioritize exome capture space #TODO: implement and test once examples are available}

\item{mrna_collapse_redundancy}{Default: TRUE. Set to FALSE to obtain all rows for the mrna seq_type including those that would otherwise be collapsed.}

\item{collapse_redundancy}{Default: TRUE. Set to FALSE to obtain all rows (except bam_available == FALSE) for all seq_types including those that would otherwise be collapsed.}

\item{also_normals}{Set to TRUE to force the return of rows where tissue_status is normal (default is to restrict to tumour)}

\item{include_unavailable}{Set to TRUE to include samples with \code{bam_available == FALSE}. Default: FALSE - only samples with \code{bam_available = TRUE} are retained.}

\item{verbose}{Set to TRUE for a chatty output (mostly for debugging)}

\item{invert}{Set to TRUE to force the function to return only the rows that are lost in all the prioritization steps (mostly for debugging)}

\item{exclude}{Specify one or more seq_type to drop from the output.}

\item{min_corrected_cov}{Minimum value for the MeanCorrectedCoverage QC
metric for the genome and capture samples. The samples with
MeanCorrectedCoverage below this value will be excluded from the
metadata output. A numeric value is expected. The default is set to 25.
This prevents metadata from containing anythong other than the three standard
seq_type (genome, capture, mrna). Default setting will exclude "promethION".}

\item{...}{Additional arguments}
}
\value{
A data frame with metadata for each biopsy in GAMBL

\describe{
\item{compression}{Format of the original data used as input for our analysis pipelines (cram, bam or fastq)}
\item{bam_available}{Whether or not this file was available when last checked.}
\item{patient_id}{The anonymized unique identifier for this patient. For BC samples, this will be Res ID.}
\item{sample_id}{A unique identifier for the sample analyzed.}
\item{seq_type}{The assay type used to produce this data (one of "genome","capture, "mrna", "promethION")}
\item{capture_space}{Unique ID for the capture space, where applicable}
\item{genome_build}{The name of the genome reference the data were aligned to.}
\item{tissue_status}{Whether the sample was atumour or normal.}
\item{cohort}{Name for a group of samples that were added together (usually from a single study), often in the format {pathology}_{cohort_descriptor}.}
\item{library_id}{The unique identifier for the sequencing library.}
\item{pathology}{The diagnosis or pathology for the sample}
\item{time_point}{Timing of biopsy in increasing alphabetical order (A = diagnosis, B = first relapse etc)}
\item{protocol}{General protocol for library construction. e.g. "Ribodepletion", "PolyA", or "Genome"}
\item{ffpe_or_frozen}{Whether the nucleic acids were extracted from a frozen or FFPE sample}
\item{read_length}{The length of reads (required for RNA-seq libraries)}
\item{strandedness}{Whether the RNA-seq librayr construction was strand-specific and, if so, which strand. Required for RNAseq; "positive", "negative", or "unstranded")}
\item{seq_source_type}{Required for RNAseq. Usually the same value as ffpe_or_frozen but sometimes immunotube or sorted cells}
\item{data_path}{Symbolic link to the bam or cram file (not usually relevant for GAMBLR)}
\item{link_name}{Standardized naming for symbolic link (not usually relevant for GAMBLR)}
\item{fastq_data_path}{Symbolic link to the fastq file (not usually relevant for GAMBLR)}
\item{fastq_link_name}{Standardized naming for symbolic link for FASTQ file, if used (not usually relevant for GAMBLR)}
\item{unix_group}{Whether a source is external and restricted by data access agreements (icgc_dart) or internal (gambl)}
\item{COO_consensus}{TODO}
\item{DHITsig_consensus}{TODO}
\item{COO_PRPS_class}{TODO}
\item{DHITsig_PRPS_class}{TODO}
\item{DLBCL90_dlbcl_call}{TODO}
\item{DLBCL90_dhitsig_call}{TODO}
\item{res_id}{duplicate of sample_id for local samples and NA otherwise}
\item{DLBCL90_code_set}{Code set used for DLBCL90 call. One of DLBCL90 DLBCL90v2 DLBCL90v3}
\item{DLBCL90_dlbcl_score}{TODO}
\item{DLBCL90_pmbl_score}{TODO}
\item{DLBCL90_pmbl_call}{TODO}
\item{DLBCL90_dhitsig_score}{TODO}
\item{myc_ba}{Result from breakapart FISH for MYC locus}
\item{myc_cn}{Result from copy number FISH for MYC locus}
\item{bcl2_ba}{Result from breakapart FISH for BCL2 locus}
\item{bcl2_cn}{Result from copy number FISH for BCL2 locus}
\item{bcl6_ba}{Result from breakapart FISH for BCL6 locus}
\item{bcl6_cn}{Result from copy number FISH for BCL6 locus}
\item{time_since_diagnosis_years}{TODO}
\item{relapse_timing}{TODO}
\item{dtbx}{TODO. OR REMOVE?}
\item{dtdx}{TODO. OR REMOVE?}
\item{lymphgen_no_cnv}{TODO}
\item{lymphgen_with_cnv}{TODO}
\item{lymphgen_cnv_noA53}{TODO}
\item{lymphgen_wright}{The LymphGen call for this sample from Wright et all (if applicable)}
\item{fl_grade}{TODO}
\item{transformation}{TODO}
\item{relapse}{TODO}
\item{ighv_mutation_original}{TODO}
}
}
\description{
Return metadata for a selection of samples.
}
\details{
This function returns metadata for GAMBL samples.
This replaces the functionality of the original version, which is still available under the new name \link{og_get_gambl_metadata}.
The purpose of this function is to provide the metadata for the non-redundant set of samples from GAMBL, dealing with all
types of redundancy caused by samples or biopsies that have data from >1 seq_type (genome or capture), different capture protocols (exome or targeted capture) etc.
}
\examples{
#basic usage
my_metadata = suppressMessages(get_gambl_metadata())
dplyr::group_by(my_metadata,pathology,seq_type) \%>\% dplyr::count()
\dontrun{
  # Rarely needed but can be useful for some applications:
  # override default filters and request metadata for samples
  # other than tumour genomes,
  #e.g. also get the normals
  tumour_and_normal_metadata = get_gambl_metadata(also_normals = TRUE))

  # prioritize exome results over genome results
  non_redundant_genome_and_capture = get_gambl_metadata(dna_seq_type_priority = "capture")
}
}
