% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fusions.R
\name{get_gene_fusions}
\alias{get_gene_fusions}
\title{Get gene fusions}
\usage{
get_gene_fusions(
  projection = "grch37",
  verbose = F,
  keep_genes,
  drop_genes,
  remove_adjacent_pairs = FALSE,
  these_samples_metadata,
  join_with_metadata = FALSE,
  harmonize_immunoglobulin_partners = FALSE
)
}
\arguments{
\item{projection}{Which genome build}

\item{verbose}{How chatty you want this experience}

\item{keep_genes}{Vector of genes to restrict to}

\item{drop_genes}{Vector of genes to drop (i.e. bad/artifacts)}

\item{remove_adjacent_pairs}{Set to TRUE to drop genes that are already in close proximity}

\item{these_samples_metadata}{The metadata to link to your fusion table}

\item{join_with_metadata}{Set to TRUE to get fusions along with the sample metadata in one data frame}

\item{harmonize_immunoglobulin_partners}{Attempt to clean up IG names so they are all consistently one of IGH, IGK or IGL}
}
\value{
A data frame in a bedpe-like format
}
\description{
Retrieve fusions identified by FusionCatcher
}
\examples{

all_fusions = get_gene_fusions()
onco_fusions = get_gene_fusions(keep_genes = c("BCL2","MYC","BCL6"))
print(head(onco_fusions))

}
