% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ssm_by_genes.R
\name{get_ssm_by_genes}
\alias{get_ssm_by_genes}
\title{Get SSM By Gene}
\usage{
get_ssm_by_genes(
  genes,
  these_samples_metadata = NULL,
  projection = "grch37",
  ashm_regions,
  drop_silent_outside_ashm_regions = FALSE,
  expand_by = 0
)
}
\arguments{
\item{genes}{Gene symbol for one or more genes}

\item{these_samples_metadata}{Optional, a metadata table (with sample IDs in a column) to subset the return to.}

\item{projection}{Obtain variants projected to this reference (one of grch37 or hg38).}
}
\value{
A data frame containing all the MAF data columns (one row per mutation).
}
\description{
Retrieve all SSMs from GAMBL for one or more genes
}
