% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timed_mutations.R
\name{get_timed_mutations}
\alias{get_timed_mutations}
\title{Load MutationTimeR Results}
\usage{
get_timed_mutations(this_sample_metadata, genome_build, verbose = FALSE)
}
\arguments{
\item{this_sample_metadata}{Metadata with one row
containing the details for the desired genome sample_id}

\item{verbose}{Set to TRUE for a chattier experience. Default is FALSE.}

\item{projection}{Genome build projection (e.g., "hg38" or "grch37").}
}
\value{
a named list containing two data.frames with
the ssm and cna timing information
}
\description{
This function retrieves and loads the MutationTimeR
timing estimates for a single sample_id for all successfully timed
and un-timed SSMs and CNAs.
}
