% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf_to_custom_track.R
\name{maf_to_custom_track}
\alias{maf_to_custom_track}
\title{Maf To Custom Track.}
\usage{
maf_to_custom_track(
  maf_data,
  these_samples_metadata = NULL,
  this_seq_type = "genome",
  output_file,
  as_bigbed = FALSE,
  as_biglolly = FALSE,
  colour_column = "lymphgen",
  track_name = "GAMBL mutations",
  track_description = "GAMBL mutations",
  verbose = FALSE,
  padding_size = 0,
  projection = "grch37",
  bedToBigBed_path = "config",
  these_sample_ids = NULL
)
}
\arguments{
\item{maf_data}{MAF data obtained from of the \code{get_ssm} family of functions.}

\item{these_samples_metadata}{A metadata table to subset the samples of interest from
the input \code{maf_data}. If NULL (the default), all samples in \code{maf_data} are kept.}

\item{this_seq_type}{The seq type used to filter the samples if \code{these_samples_metadata}
is not provided. Should match what was used to get \code{maf_data}. Default is "genome".}

\item{output_file}{Name for your new bed file that can be uploaded as a custom track to UCSC.}

\item{as_bigbed}{Boolean parameter controlling the format of the created track file.
Default is FALSE.}

\item{as_biglolly}{Boolean parameter controlling the format of the created track file.
Default is FALSE (i.e a BED file will be returned).}

\item{colour_column}{Set the colouring the SSMs in the track. Possible options are lymphgen
(default), cohort, pathology, genome_build (as in \code{these_samples_metadata}), and mutation
(corresponds to MAF Variant_Classification).}

\item{track_name}{Track name to use in the header if output is not bigBed or bigLolly.
Default is "GAMBL mutations"}

\item{track_description}{Track description to use in the header if output is not
bigBed or bigLolly. Default is "GAMBL mutations"}

\item{verbose}{Default is FALSE.}

\item{padding_size}{Optional parameter specifying the padding size in the
returned file, default is 0.}

\item{projection}{Specify which genome build to use. Possible values are "grch37" (default)
or "hg38". This parameter has an effect only when \code{as_bigbed} or \code{as_biglolly} is TRUE.}

\item{bedToBigBed_path}{Path to your local \code{bedToBigBed} UCSC tool or the string
\code{"config"} (default). If set to \code{"config"}, \code{GAMBLR.helpers::check_config_value}
is called internally and the \code{bedToBigBed} path is obtained from the \code{config.yml}
file saved in the current working directory. This parameter is ignored if both
\code{as_bigbed} and \code{as_biglolly} is set to \code{FALSE}.}

\item{these_sample_ids}{DEPRECATED}
}
\value{
Nothing.
}
\description{
Convert mutations into a UCSC custom track file
}
\details{
This function takes a set of mutations as maf_data and converts it
to a UCSC Genome Browser ready BED (or bigbed/biglolly) file complete with
the required header. Upload the resulting file to
\href{https://genome.ucsc.edu/cgi-bin/hgCustom}{UCSC genome browser}
to view your data as a custom track.
Optional parameters available for further customization of the returned file.
For more information, refer to the parameter descriptions
and function examples.
}
\examples{
# using grch37 coordinates
myc_grch37 <- GAMBLR.utils::create_bed_data(
                GAMBLR.data::grch37_lymphoma_genes_bed
              ) \%>\%
              dplyr::filter(name == "MYC")

print(myc_grch37)
# desired projection will be automatically set to the
# genome_build of your region object
genome_maf <- get_ssm_by_regions(regions_bed = myc_grch37,
                             these_samples_metadata = get_gambl_metadata(),
                             this_seq_type = "genome",
                             streamlined = FALSE,
                             basic_columns = TRUE)

# myc_hg19.bed will be created in your working directory
maf_to_custom_track(maf_data = genome_maf, output_file = "myc_genome_hg19.bed")

#lazy/concise way:
my_region = "8:128747680-128753674"

capture_maf <- get_ssm_by_regions(regions_list = my_region,
                             these_samples_metadata = get_gambl_metadata(),
                             this_seq_type = "capture",
                             projection = "grch37",
                             streamlined = FALSE,
                             basic_columns = TRUE)
maf_to_custom_track(maf_data = capture_maf, this_seq_type = "capture", output_file = "myc_capture_hg19.bed")

}
